requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.9.0
attrs==21.2.0
beautifulsoup4==4.10.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
ConfigArgParse==1.5.3
dataclasses==0.8
Django==3.0.1
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-ipware==4.0.2
django-livereload-server==0.4
django-prometheus==2.2.0
django-redis==5.1.0
django-socketio==0.2
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.1.0
execnet==1.9.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
gevent-socketio==0.3.6
gevent-websocket==0.10.1
geventhttpclient==1.5.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
locust==2.5.1
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.3
mypy==0.930
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
parso==0.8.3
platformdirs==2.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.8.0
psycopg2==2.8.6
ptpython==3.0.20
py==1.11.0
Pygments==2.10.0
pylint==2.12.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymongo==3.12.1
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
pytz==2021.3
pyzmq==22.3.0
redis==3.5.3
requests==2.26.0
rich==10.16.1
roundrobin==0.0.2
sentry-sdk==1.5.1
six==1.16.0
soupsieve==2.3.1
sphinx-me==0.3
sqlparse==0.2.4
toml==0.10.2
tomli==1.2.3
tornado==6.1
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
